+++ state.c
@@ -2,9 +2,6 @@
 // state.c
 //
 
-#include <pic18f26k22.h>
-#include <xc.h>
-
 #include "state.h"
 #include "emc_library.h"
 #include "io.h"
@@ -126,7 +123,7 @@ uint8_t check_supervisory()
             {
                 if( ++device_id_table[cnt1][7] >= 84 )
                 {            
-                    add_event(SUPERVISORY_T,cnt1+3);
+                    add_event(SUPERVISORY_T,cnt1+3U);
                     device_id_table[cnt1][7] = 'S';
                 }
             }
@@ -158,7 +155,7 @@ void process_ADC()
 		
         // ADRESH: ADC RESULT REGISTER HIGH (ADRESH) ADFM = 1
         // ADRESL: ADC RESULT REGISTER LOW (ADRESL) ADFM = 1
-		ADC_data = (ADRESH<<8)+ADRESL;
+		ADC_data = (uint16_t) ((ADRESH<<8U)+ADRESL);
 		ADC_data &= 0x03ff;
 		ADON = 0;                               //748->2.80 V
 		ADC_time = 0;                           //776->2.70 V
